Cleanup code, moving DependentData inside a static class wrap external calls.
