+++ src/Microsoft.VisualStudio.Threading/IJoinableTaskDependent.cs
@@ -24,7 +24,7 @@ namespace Microsoft.VisualStudio.Threading
         /// <summary>
         /// Get the reference of dependent node to record dependencies.
         /// </summary>
-        ref JoinableTaskDependentData GetJoinableTaskDependentData();
+        ref JoinableTaskDependencyGraph.JoinableTaskDependentData GetJoinableTaskDependentData();
 
         /// <summary>
         /// A function is called, when this dependent node is added to be a dependency of a parent node.
