+++ src/Microsoft.VisualStudio.Threading/JoinableTaskFactory.cs
@@ -376,7 +376,7 @@ namespace Microsoft.VisualStudio.Threading
                     var allJoinedJobs = new HashSet<JoinableTask>();
                     lock (this.Context.SyncContextLock)
                     {
-                        currentBlockingTask.GetJoinableTaskDependentData().AddSelfAndDescendentOrJoinedJobs(allJoinedJobs);
+                        JoinableTaskDependencyGraph.AddSelfAndDescendentOrJoinedJobs(currentBlockingTask, allJoinedJobs);
                         return allJoinedJobs.Any(t => (t.CreationOptions & JoinableTaskCreationOptions.LongRunning) == JoinableTaskCreationOptions.LongRunning);
                     }
                 }
@@ -972,7 +972,7 @@ namespace Microsoft.VisualStudio.Threading
                 // Join the ambient parent job, so the parent can dequeue this job's work.
                 if (this.previousJoinable != null && !this.previousJoinable.IsCompleted)
                 {
-                    this.previousJoinable.GetJoinableTaskDependentData().AddDependency(joinable);
+                    JoinableTaskDependencyGraph.AddDependency(this.previousJoinable, joinable);
 
                     // By definition we inherit the nesting factories of our immediate nesting task.
                     var nestingFactories = this.previousJoinable.NestingFactories;
